move datatoken name generation out of class
