+++ src/datatokens/Datatokens.ts
@@ -4,10 +4,10 @@ import { AbiItem } from 'web3-utils/types'
 import defaultFactoryABI from '@oceanprotocol/contracts/artifacts/DTFactory.json'
 import defaultDatatokensABI from '@oceanprotocol/contracts/artifacts/DataTokenTemplate.json'
 import { Logger, getFairGasPrice } from '../utils'
-import wordListDefault from '../data/words.json'
 import { TransactionReceipt } from 'web3-core'
 import BigNumber from 'bignumber.js'
 import Decimal from 'decimal.js'
+import { generateDatatokenName } from '../utils/Datatokens'
 
 /**
  * Provides an interface to DataTokens
@@ -51,21 +51,7 @@ export class DataTokens {
     name: string
     symbol: string
   } {
-    const list = wordList || wordListDefault
-    const random1 = Math.floor(Math.random() * list.adjectives.length)
-    const random2 = Math.floor(Math.random() * list.nouns.length)
-    const indexNumber = Math.floor(Math.random() * 100)
-
-    // Capitalized adjective & noun
-    const adjective = list.adjectives[random1].replace(/^\w/, (c) => c.toUpperCase())
-    const noun = list.nouns[random2].replace(/^\w/, (c) => c.toUpperCase())
-
-    const name = `${adjective} ${noun} Token`
-    // use first 3 letters of name, uppercase it, and add random number
-    const symbol = `${(
-      adjective.substring(0, 3) + noun.substring(0, 3)
-    ).toUpperCase()}-${indexNumber}`
-
+    const { name, symbol } = generateDatatokenName(wordList)
     return { name, symbol }
   }
