+++ src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './SubscribablePromise'
 export * from './SubscribableObserver'
 export * from './GasUtils'
 export * from './AssetResolverHelper'
+export * from './Datatokens'
